requirements.txt
discord.py==2.3.2
requests
flask